Zmena penazi na big decimal vsade

+++ src/main/java/com/dejvo/Shop/db/crudservice/CustomerAccountInterface.java
@@ -4,6 +4,7 @@ package com.dejvo.Shop.db.crudservice;
 import com.dejvo.Shop.model.Customer;
 import com.dejvo.Shop.model.CustomerAccount;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -12,6 +13,6 @@ public interface CustomerAccountInterface {
     public Integer createCustomerAccount(CustomerAccount customerAccount);
     public CustomerAccount getCustomerByIdOfCustomer(int idofcustomer);
     public List<CustomerAccount> getAllCustomerAccounts();
-    public Integer updateMoneyOfCustomerAccount(double money, int idofcustomer);
-    public double getMoneyByCustomerId(int idofcustomer);
+    public Integer updateMoneyOfCustomerAccount(BigDecimal money, int idofcustomer);
+    public BigDecimal getMoneyByCustomerId(int idofcustomer);
 }
